requirements.txt
aiohttp==3.11.16
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
colormath==3.0.0
discord.py==2.5.0
frozenlist==1.3.1
idna==3.7
multidict==6.0.2
networkx==3.1
numpy==1.25.1
yarl==1.19.0
Pillow==10.3.0
requests==2.32.0
tzdata==2025.2